# specs/001-taboot-rag-platform/contracts/api-openapi.yaml
openapi: 3.0.3
info:
  title: Taboot API
  description: Doc-to-Graph RAG platform API for ingestion, extraction, and retrieval
  version: 1.0.0
  contact:
    name: Taboot Project
    url: https://github.com/yourusername/taboot
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: http://taboot-app:8000
    description: Docker container (internal)
...
security:
  - bearerAuth: []
paths:
...
  /ingest:
    post:
      tags: [ingestion]
      summary: Start an ingestion job
      description: Ingest documents from specified source with optional limit
      operationId: start_ingestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestionRequest'
      responses:
        '202':
          description: Ingestion job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJobResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /extract/pending:
    post:
      tags: [extraction]
      summary: Process documents awaiting extraction
      description: Start extraction worker to process documents in pending state
      operationId: extract_pending
      responses:
        '202':
          description: Extraction started
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [started]
                  documents_queued:
                    type: integer
                    example: 42
        '500':
          description: Extraction failed to start
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /status:
    get:
      tags: [status]
      summary: Get system health status
      description: Check health of all services and report configuration
      operationId: get_status
      parameters:
        - name: component
          in: query
          required: false
          schema:
            type: string
            enum: [neo4j, qdrant, redis, tei, ollama, firecrawl, playwright]
        - name: verbose
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemStatus'
  /documents:
    get:
      tags: [status]
      summary: List ingested documents
      description: Retrieve paginated list of documents with optional filters
      operationId: list_documents
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: source_type
          in: query
          required: false
          schema:
            type: string
            enum: [web, github, reddit, youtube, gmail, elasticsearch, docker_compose, swag, tailscale, unifi, ai_session]
        - name: extraction_state
          in: query
          required: false
          schema:
            type: string
            enum: [pending, tier_a_done, tier_b_done, tier_c_done, completed, failed]
      responses:
        '200':
          description: Documents retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Document'
                  total:
                    type: integer
                    example: 150
                  limit:
                    type: integer
                    example: 20
                  offset:
                    type: integer
                    example: 0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token authentication for API access
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid source type"
        details:
          type: object
          additionalProperties: true
    IngestionRequest:
      type: object
      required:
        - source_type
        - source_target
      properties:
        source_type:
          type: string
          enum: [web, github, reddit, youtube, gmail, elasticsearch, docker_compose, swag, tailscale, unifi, ai_session]
          example: web
        source_target:
          type: string
          description: URL, repo name, file path, or other source identifier
          example: "https://docs.example.com"
        limit:
          type: integer
          description: Maximum pages/items to ingest (optional, defaults to unlimited)
          minimum: 1
          example: 20
    IngestionJobResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        state:
          type: string
          enum: [pending, running, completed, failed]
          example: pending
        source_type:
          type: string
          example: web
        source_target:
          type: string
          example: "https://docs.example.com"
        created_at:
          type: string
          format: date-time
          example: "2025-10-21T12:34:56Z"
...
    SystemStatus:
      type: object
      properties:
        services:
          type: object
          additionalProperties:
            type: object
            properties:
              status:
                type: string
                enum: [healthy, unhealthy]
              message:
                type: string
                nullable: true
          example:
            neo4j:
              status: healthy
              message: null
            qdrant:
              status: healthy
              message: null
            redis:
              status: healthy
              message: null
        queue_depth:
          type: object
          properties:
            ingestion:
              type: integer
              example: 0
            extraction:
              type: integer
              example: 5
        config:
          type: object
          additionalProperties:
            type: string
          example:
            neo4j_uri: "bolt://taboot-graph:7687"
            qdrant_url: "http://taboot-vectors:6333"
            embedding_model: "Qwen3-Embedding-0.6B"
    Document:
      type: object
      properties:
        doc_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        source_url:
          type: string
          example: "https://docs.example.com/page"
        source_type:
          type: string
          enum: [web, github, reddit, youtube, gmail, elasticsearch, docker_compose, swag, tailscale, unifi, ai_session]
          example: web
        content_hash:
          type: string
          pattern: '^[a-f0-9]{64}$'
          example: "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
        ingested_at:
          type: string
          format: date-time
          example: "2025-10-21T12:00:00Z"
        extraction_state:
          type: string
          enum: [pending, tier_a_done, tier_b_done, tier_c_done, completed, failed]
          example: completed
        extraction_version:
          type: string
          nullable: true
          example: "v1.0.0"
        updated_at:
          type: string
          format: date-time
          example: "2025-10-21T12:05:30Z"